backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
reportlab==4.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
sendgrid==6.11.0
pytz==2024.2
celery==5.4.0
redis==5.2.0
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
email-validator==2.2.0
